+++ src/main/java/com/capstone/cschatbot/cs/application/evaluation/EvaluationService.java
@@ -1,6 +1,6 @@
 package com.capstone.cschatbot.cs.application.evaluation;
 
-import com.capstone.cschatbot.cs.domain.ChatEvaluation;
+import com.capstone.cschatbot.cs.dto.ChatEvaluation;
 
 import java.util.concurrent.CompletableFuture;
