+++ src/main/java/com/capstone/cschatbot/cs/dto/response/CSChatDto.java
@@ -1,6 +1,6 @@
 package com.capstone.cschatbot.cs.dto.response;
 
-import com.capstone.cschatbot.cs.domain.ChatEvaluation;
+import com.capstone.cschatbot.cs.dto.ChatEvaluation;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.fasterxml.jackson.datatype.jsr310.deser.LocalDateDeserializer;
